Fix code flagged by stricter warnings
Narrowing conversions (-Wconversion), float to double promotions
(-Wdouble-promotion) and a local shadowing another local
(-Wshadow=local). Unit conversions now go through std::chrono instead
of hand-written factors, and getAddress() returns the uint16_t the I2C
driver stores.

+++ components/kernel/src/I2CManager.hpp
@@ -128,7 +128,7 @@ public:
         return bus;
     }
 
-    uint8_t getAddress() const {
+    uint16_t getAddress() const {
         return device.addr;
     }
